separate module for csv field info
